requirements.txt
streamlit
streamlit-mic-recorder
deepface
openai-whisper
transformers
torch
tensorflow
opencv-python
numpy